\name{DiversitySampler-Package}
\alias{DiversitySampler-Package}
\alias{DiversitySampler}
\docType{package}
\title{
Re-Sampling to Estimate Diversity Indices
}
\description{
Contains functions for conducting analyses of expected diversity or evenness at a given sampling level. A Monte-Carlo procedure is used to re-sample a given observation at each level of sampling. The expectation being that the mean of the re-sampling will approach Shannon's diversity index at that sample level.
}
\details{
\tabular{ll}{
Package: \tab DiversitySampler\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2009-12-07\cr
License: \tab GPL 3 or newer \cr
LazyLoad: \tab yes\cr
}

}
\author{

Maintainer: Matthew K. Lau <mkl48@nau.edu>, Department of Biological Sciences, Northern Arizona University, AZ, USA 

Raj Whitlock, Department of Animal and Plant Sciences, University of Sheffield, Sheffield, UK


}
%% \references{

%% }
\keyword{iteration}
\keyword{datagen}

\seealso{\code{\link{sample}}
}
