\name{DiversitySampler-package}
\alias{DiversitySampler-package}
\alias{DiversitySampler}
\docType{package}
\title{
Re-Sampling to Estimate Shannon's Diversity Index
}
\description{
There are two functions in this package, which can be used together to estimate the Shannon's Diversity index at different levels of sample size. A Monte-Carlo procedure is used to re-sample a given observation at each level of sampling. The expectation being that the mean of the re-sampling will approach Shannon's diversity index at that sample level.
}
\details{
\tabular{ll}{
Package: \tab DiversitySampler\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-07-23\cr
License: \tab GPL 2 or newer \cr
LazyLoad: \tab yes\cr
}

}
\author{
Matthew K. Lau

Maintainer: <mkl48@nau.edu>

}
\references{

}
\keyword{iteration}
\keyword{datagen}

\seealso{
}
\examples{

}
